requirements.txt
imgbbpy==0.1.3
loguru==0.6.0
platformdirs==3.1.0
PyMDL==1.0.1
pymediainfo==6.0.1
pyperclip==1.8.2
PyYAML==6.0
requests==2.28.2
sv-ttk==2.4